+++ src/main/java/akkimi_BE/aja/dto/auth/KakaoLoginRequest.java
@@ -1,4 +1,4 @@
-package akkimi_BE.aja.dto.oauth;
+package akkimi_BE.aja.dto.auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
