--- src/main/java/akkimi_BE/aja/dto/request/SocialSignupRequestDto.java
@@ -1,9 +1,0 @@
-package akkimi_BE.aja.dto.request;
-
-import akkimi_BE.aja.entity.SocialType;
-
-public record SocialSignupRequestDto(
-        SocialType socialType,
-        String socialId,
-        String nickname
-) {}
