+++ src/main/java/akkimi_BE/aja/dto/request/UpdateRegionRequestDto.java
@@ -3,6 +3,6 @@ package akkimi_BE.aja.dto.request;
 import lombok.Getter;
 
 @Getter
-public class UpdateRegionRequest {
+public class UpdateRegionRequestDto {
     private String region;
 }
