+++ src/main/java/akkimi_BE/aja/global/config/AiConfig.java
@@ -1,4 +1,4 @@
-package global.config;
+package akkimi_BE.aja.global.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
