+++ src/main/java/akkimi_BE/aja/global/config/SecurityConfig.java
@@ -1,6 +1,6 @@
-package global.config;
+package akkimi_BE.aja.global.config;
 
-import global.JwtTokenFilter;
+import akkimi_BE.aja.global.JwtTokenFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
