+++ src/main/java/akkimi_BE/aja/global/exception/CustomException.java
@@ -1,4 +1,4 @@
-package global.exception;
+package akkimi_BE.aja.global.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
