+++ src/main/java/akkimi_BE/aja/global/util/TokenType.java
@@ -1,4 +1,4 @@
-package global.util;
+package akkimi_BE.aja.global.util;
 
 public enum TokenType {
     ACCESS, REFRESH
