+++ src/main/java/akkimi_BE/aja/repository/MaltuRepository.java
@@ -5,7 +5,6 @@ import akkimi_BE.aja.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Collection;
 import java.util.List;
 
 @Repository
