+++ src/main/java/akkimi_BE/aja/service/MaltuService.java
@@ -6,8 +6,8 @@ import akkimi_BE.aja.dto.response.MaltuResponseDto;
 import akkimi_BE.aja.entity.Maltu;
 import akkimi_BE.aja.entity.User;
 import akkimi_BE.aja.repository.MaltuRepository;
-import global.exception.CustomException;
-import global.exception.HttpErrorCode;
+import akkimi_BE.aja.global.exception.CustomException;
+import akkimi_BE.aja.global.exception.HttpErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
