+++ src/main/java/akkimi_BE/aja/service/UserService.java
@@ -2,7 +2,7 @@ package akkimi_BE.aja.service;
 
 import akkimi_BE.aja.dto.request.*;
 import akkimi_BE.aja.dto.response.CurrentMaltuResponseDto;
-import akkimi_BE.aja.dto.response.TokenResponse;
+import akkimi_BE.aja.dto.auth.TokenResponse;
 import akkimi_BE.aja.dto.response.UserProfileResponseDto;
 import akkimi_BE.aja.entity.Maltu;
 import akkimi_BE.aja.entity.Role;
@@ -11,9 +11,9 @@ import akkimi_BE.aja.entity.User;
 import akkimi_BE.aja.repository.MaltuRepository;
 import akkimi_BE.aja.repository.UserRepository;
 import akkimi_BE.aja.service.auth.RefreshTokenService;
-import global.exception.CustomException;
-import global.exception.HttpErrorCode;
-import global.util.JwtUtil;
+import akkimi_BE.aja.global.exception.CustomException;
+import akkimi_BE.aja.global.exception.HttpErrorCode;
+import akkimi_BE.aja.global.util.JwtUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
